Create Current bill(E)
Write a program to calculate and print the Electricity bill of a given customer. Units consumed by the user should be taken from the keyboard and display the total amount to be paid by the customer. The charges are as follows
Unit Charge/unit
upto 199 @1.20
200 and above but less than 400 @1.50
400 and above but less than 600 @1.80
600 and above @2.00
If the bill exceeds Rs. 400 then a surcharge of 15% will be charged.
If the bill is less than Rs. 400 then a minimum surcharge amount should be Rs. 100/-.